# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApoTrack API",
        "description": "Dokumentasi API resmi untuk platform ApoTrack — sistem manajemen apotek digital. API ini digunakan oleh aplikasi mobile Flutter untuk Customer dan Staff Apotek.",
        "contact": {
            "name": "Tim Pengembang ApoTrack",
            "email": "dev@apotrack.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server Aktif (Local Development)"
        }
    ],
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Lihat isi keranjang belanja Customer",
                "description": "Mengambil seluruh item obat yang ada di keranjang belanja pengguna saat ini beserta total harganya. Memerlukan token Sanctum.",
                "operationId": "getCart",
                "responses": {
                    "200": {
                        "description": "Data keranjang berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data keranjang berhasil diambil"
                                        },
                                        "data": {
                                            "properties": {
                                                "pharmacy": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid",
                                                            "example": "01932c91-5f00-7341-91a3-b9e7f02c8d1a"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Apotek Sehat Bahagia"
                                                        },
                                                        "address": {
                                                            "type": "string",
                                                            "example": "Jl. Merdeka No. 123"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "format": "uuid",
                                                                "example": "019e2ac0-684e-7139-8cb9-c150e027c872"
                                                            },
                                                            "medicine_id": {
                                                                "type": "string",
                                                                "format": "uuid",
                                                                "example": "019e2ac0-684e-7139-8cb9-c150e027c872"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Paracetamol 500mg"
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "price": {
                                                                "type": "integer",
                                                                "example": 5000
                                                            },
                                                            "subtotal": {
                                                                "type": "integer",
                                                                "example": 10000
                                                            },
                                                            "image_url": {
                                                                "type": "string",
                                                                "example": "https://example.com/medicine.jpg"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total_price": {
                                                    "type": "integer",
                                                    "example": 10000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cart/items": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Tambahkan item obat ke keranjang belanja",
                "description": "Menambahkan obat ke dalam keranjang. Jika keranjang sudah berisi obat dari apotek lain, sistem akan memvalidasi atau mengganti keranjang sesuai aturan bisnis.",
                "operationId": "addCartItem",
                "requestBody": {
                    "description": "Data item obat yang ditambahkan",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pharmacy_id",
                                    "medicine_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "pharmacy_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "01932c91-5f00-7341-91a3-b9e7f02c8d1a"
                                    },
                                    "medicine_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "019e2ac0-684e-7139-8cb9-c150e027c872"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2,
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item berhasil ditambahkan ke keranjang"
                                        },
                                        "data": {
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal (stok tidak cukup atau ID tidak valid)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Stok obat tidak mencukupi."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cart/items/{id}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Hapus item dari keranjang belanja",
                "description": "Menghapus item obat tertentu dari keranjang berdasarkan ID item keranjang (CartItem ID).",
                "operationId": "removeCartItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID item keranjang (CartItem UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "019e2ac0-684e-7139-8cb9-c150e027c872"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item berhasil dihapus dari keranjang"
                                        },
                                        "data": {
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item tidak ditemukan di keranjang",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item keranjang tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum Anda di sini. Token didapatkan dari endpoint POST /api/auth/login. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}